+++ src/index.html
@@ -18,7 +18,7 @@
             <ul class="scores-display">
             </ul>
         </div>
-        <form action="#" class="add-score">
+        <form action="#" class="add-score" id="form">
             <h3 class="add-heading">Add your score</h3>
             <input class="form-input" id="name" type="text" placeholder="Your Name">
             <input class="form-input" id="score" type="number" placeholder="Your Score">
